Setup CI build (#11)
Add a .yml configuration that runs CI builds on the robot code.
This will auto-compile the code and check for errors every time 
there is a push to master or a pull request opened where the 
destination branch is master.  This should prevent team members 
from accidentally committing code with errors that may break master.